# Lift-Simple/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lift Simple API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserRegistration {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export type PlanUpdatePlan = typeof PlanUpdatePlan[keyof typeof PlanUpdatePlan];


export const PlanUpdatePlan = {
  free: 'free',
  premium: 'premium',
} as const;

export interface PlanUpdate {
  plan: PlanUpdatePlan;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  premium: 'premium',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  plan: UserPlan;
  createdAt: string;
}

export interface AuthResult {
  user: User;
}

export type ExerciseCategory = typeof ExerciseCategory[keyof typeof ExerciseCategory];


export const ExerciseCategory = {
  lower_body: 'lower_body',
  upper_body: 'upper_body',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  full_body: 'full_body',
} as const;

export type ExerciseUnit = typeof ExerciseUnit[keyof typeof ExerciseUnit];


export const ExerciseUnit = {
  lbs: 'lbs',
  kg: 'kg',
} as const;

export type MovementType = typeof MovementType[keyof typeof MovementType];


export const MovementType = {
  upper_body: 'upper_body',
  lower_body: 'lower_body',
} as const;

export interface Exercise {
  id: number;
  name: string;
  category: ExerciseCategory;
  /** @nullable */
  notes?: string | null;
  unit: ExerciseUnit;
  isCustom: boolean;
  /** @nullable */
  userId?: number | null;
  targetRepMin?: number;
  targetRepMax?: number;
  movementType: MovementType;
}

export type ExerciseInputCategory = typeof ExerciseInputCategory[keyof typeof ExerciseInputCategory];


export const ExerciseInputCategory = {
  lower_body: 'lower_body',
  upper_body: 'upper_body',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  full_body: 'full_body',
} as const;

export type ExerciseInputUnit = typeof ExerciseInputUnit[keyof typeof ExerciseInputUnit];


export const ExerciseInputUnit = {
  lbs: 'lbs',
  kg: 'kg',
} as const;

export interface ExerciseInput {
  /** @minLength 1 */
  name: string;
  category: ExerciseInputCategory;
  notes?: string;
  unit?: ExerciseInputUnit;
  targetRepMin?: number;
  targetRepMax?: number;
  movementType: MovementType;
}

export type ExerciseUpdateCategory = typeof ExerciseUpdateCategory[keyof typeof ExerciseUpdateCategory];


export const ExerciseUpdateCategory = {
  lower_body: 'lower_body',
  upper_body: 'upper_body',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  full_body: 'full_body',
} as const;

export type ExerciseUpdateUnit = typeof ExerciseUpdateUnit[keyof typeof ExerciseUpdateUnit];


export const ExerciseUpdateUnit = {
  lbs: 'lbs',
  kg: 'kg',
} as const;

export interface ExerciseUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ExerciseUpdateCategory;
  /** @nullable */
  notes?: string | null;
  unit?: ExerciseUpdateUnit;
  targetRepMin?: number;
  targetRepMax?: number;
  movementType?: MovementType;
}

export interface ExerciseBest {
  exerciseId: number;
  /** @nullable */
  exerciseName?: string | null;
  /** @nullable */
  bestOneRepMax?: number | null;
  /** @nullable */
  bestWeight?: number | null;
  /** @nullable */
  bestReps?: number | null;
  /** @nullable */
  achievedAt?: string | null;
}

export interface Workout {
  id: number;
  userId: number;
  /** @nullable */
  title?: string | null;
  date: string;
  isCompleted: boolean;
  createdAt: string;
}

export interface WorkoutInput {
  title?: string;
  date: string;
}

export interface WorkoutUpdate {
  /** @nullable */
  title?: string | null;
  date?: string;
}

export interface WorkoutSet {
  id: number;
  workoutExerciseId: number;
  /** @minimum 0 */
  weight: number;
  /**
     * @minimum 1
     * @maximum 9999
     */
  reps: number;
  /** @nullable */
  notes?: string | null;
  setOrder: number;
  /** @nullable */
  estimatedOneRepMax?: number | null;
  isCompleted?: boolean;
}

export interface SetInput {
  /** @minimum 0 */
  weight: number;
  /**
     * @minimum 1
     * @maximum 9999
     */
  reps: number;
  notes?: string;
  isCompleted?: boolean;
}

export interface SetUpdate {
  /** @minimum 0 */
  weight?: number;
  /**
     * @minimum 1
     * @maximum 9999
     */
  reps?: number;
  /** @nullable */
  notes?: string | null;
  isCompleted?: boolean;
}

export interface WorkoutExerciseInput {
  exerciseId: number;
}

export interface WorkoutExerciseDetail {
  id: number;
  workoutId: number;
  exerciseId: number;
  exercise: Exercise;
  sets: WorkoutSet[];
  lastTimeSets: WorkoutSet[];
}

export interface WorkoutDetail {
  id: number;
  userId: number;
  /** @nullable */
  title?: string | null;
  date: string;
  isCompleted: boolean;
  createdAt: string;
  exercises: WorkoutExerciseDetail[];
}

export interface WorkoutSummary {
  workoutId: number;
  message: string;
  highlights: string[];
  totalSets?: number;
  totalExercises?: number;
}

export interface ProgressPoint {
  date: string;
  value: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  reps?: number | null;
}

/**
 * @nullable
 */
export type ExerciseProgressTrend = typeof ExerciseProgressTrend[keyof typeof ExerciseProgressTrend] | null;


export const ExerciseProgressTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface ExerciseProgress {
  exerciseId: number;
  exerciseName: string;
  dataPoints: ProgressPoint[];
  /** @nullable */
  bestOneRepMax?: number | null;
  /** @nullable */
  trend?: ExerciseProgressTrend;
  isPremiumLocked?: boolean;
}

export interface ProgressData {
  exercises: ExerciseProgress[];
  range: string;
}

export interface DashboardStats {
  totalWorkouts: number;
  thisWeekWorkouts: number;
  currentStreak?: number;
  recentWorkouts: Workout[];
  topExercises: ExerciseBest[];
  /** @nullable */
  lastWorkoutSummary?: string | null;
}

export type GetProgressParams = {
exerciseId?: number;
range?: GetProgressRange;
};

export type GetProgressRange = typeof GetProgressRange[keyof typeof GetProgressRange];


export const GetProgressRange = {
  week: 'week',
  month: 'month',
  three_months: 'three_months',
  six_months: 'six_months',
  year: 'year',
  all: 'all',
} as const;