# Lift-Simple/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lift Simple API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/me/plan:
    patch:
      operationId: updateUserPlan
      tags: [auth]
      summary: Toggle subscription plan (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Exercises
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List all exercises (library + user custom)
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create a custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          description: Validation error
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get a single exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
    patch:
      operationId: updateExercise
      tags: [exercises]
      summary: Update a custom exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseUpdate"
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete a custom exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /exercises/{id}/best:
    get:
      operationId: getExerciseBest
      tags: [exercises]
      summary: Get best 1RM estimate and personal record for an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Best performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseBest"
        "404":
          description: Not found

  # Workouts
...
  /workout-exercises/{workoutExerciseId}/sets:
    post:
      operationId: addSet
      tags: [sets]
      summary: Add a set to a workout exercise
      parameters:
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetInput"
      responses:
        "201":
          description: Set added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
        "400":
          description: Validation error
  /sets/{id}:
    patch:
      operationId: updateSet
      tags: [sets]
      summary: Update a set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUpdate"
      responses:
        "200":
          description: Updated set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
        "404":
          description: Not found
    delete:
      operationId: deleteSet
      tags: [sets]
      summary: Delete a set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Progress
...
components:
  schemas:
...
    UserRegistration:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    PlanUpdate:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [free, premium]
    User:
      type: object
      required: [id, email, name, plan, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, premium]
        createdAt:
          type: string
          format: date-time
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # Exercise schemas
    Exercise:
      type: object
      required: [id, name, category, unit, isCustom, movementType]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [lower_body, upper_body, back, shoulders, arms, core, full_body]
        notes:
          type: ["string", "null"]
        unit:
          type: string
          enum: [lbs, kg]
        isCustom:
          type: boolean
        userId:
          type: ["integer", "null"]
        targetRepMin:
          type: integer
          default: 6
        targetRepMax:
          type: integer
          default: 12
        movementType:
          type: string
          enum: [upper_body, lower_body]
          default: upper_body
    ExerciseInput:
      type: object
      required: [name, category, movementType]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [lower_body, upper_body, back, shoulders, arms, core, full_body]
        notes:
          type: string
        unit:
          type: string
          enum: [lbs, kg]
          default: lbs
        targetRepMin:
          type: integer
          default: 6
        targetRepMax:
          type: integer
          default: 12
        movementType:
          type: string
          enum: [upper_body, lower_body]
          default: upper_body
    ExerciseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [lower_body, upper_body, back, shoulders, arms, core, full_body]
        notes:
          type: ["string", "null"]
        unit:
          type: string
          enum: [lbs, kg]
        targetRepMin:
          type: integer
        targetRepMax:
          type: integer
        movementType:
          type: string
          enum: [upper_body, lower_body]
    ExerciseBest:
      type: object
      required: [exerciseId]
      properties:
        exerciseId:
          type: integer
        exerciseName:
          type: ["string", "null"]
        bestOneRepMax:
          type: ["number", "null"]
        bestWeight:
          type: ["number", "null"]
        bestReps:
          type: ["integer", "null"]
        achievedAt:
          type: ["string", "null"]
          format: date-time

    # Workout schemas
...
    WorkoutSet:
      type: object
      required: [id, workoutExerciseId, weight, reps, setOrder]
      properties:
        id:
          type: integer
        workoutExerciseId:
          type: integer
        weight:
          type: number
          minimum: 0
        reps:
          type: integer
          minimum: 1
          maximum: 9999
        notes:
          type: ["string", "null"]
        setOrder:
          type: integer
        estimatedOneRepMax:
          type: ["number", "null"]
        isCompleted:
          type: boolean
          default: false
    SetInput:
      type: object
      required: [weight, reps]
      properties:
        weight:
          type: number
          minimum: 0
        reps:
          type: integer
          minimum: 1
          maximum: 9999
        notes:
          type: string
        isCompleted:
          type: boolean
          default: false
    SetUpdate:
      type: object
      properties:
        weight:
          type: number
          minimum: 0
        reps:
          type: integer
          minimum: 1
          maximum: 9999
        notes:
          type: ["string", "null"]
        isCompleted:
          type: boolean
...