# Lift-Simple/lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lift Simple API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseCategory } from './exerciseCategory';
import type { ExerciseUnit } from './exerciseUnit';
import type { MovementType } from './movementType';

export interface Exercise {
  id: number;
  name: string;
  category: ExerciseCategory;
  /** @nullable */
  notes?: string | null;
  unit: ExerciseUnit;
  isCustom: boolean;
  /** @nullable */
  userId?: number | null;
  targetRepMin?: number;
  targetRepMax?: number;
  movementType: MovementType;
}