# Lift-Simple/lib/api-zod/src/generated/types/workoutSet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Lift Simple API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutSet {
  id: number;
  workoutExerciseId: number;
  /** @minimum 0 */
  weight: number;
  /**
     * @minimum 1
     * @maximum 9999
     */
  reps: number;
  /** @nullable */
  notes?: string | null;
  setOrder: number;
  /** @nullable */
  estimatedOneRepMax?: number | null;
  isCompleted?: boolean;
}